+++ src/steps/account/index.ts
@@ -4,26 +4,28 @@ import {
 } from '@jupiterone/integration-sdk-core';
 
 import { IntegrationConfig } from '../../config';
-import { Steps, Entities } from '../constants';
+import { ACCOUNT_ENTITY_KEY, Entities, Steps } from '../constants';
 import { createAccountEntity } from './converter';
 
-export const ACCOUNT_ENTITY_KEY = 'entity:account';
-
-export async function fetchAccountDetails({
+export async function fetchAccount({
+  instance,
   jobState,
 }: IntegrationStepExecutionContext<IntegrationConfig>) {
-  const accountEntity = await jobState.addEntity(createAccountEntity());
-
+  const accountEntity = createAccountEntity({
+    email: instance.config.email,
+    hostname: instance.config.hostname,
+  });
+  await jobState.addEntity(accountEntity);
   await jobState.setData(ACCOUNT_ENTITY_KEY, accountEntity);
 }
 
 export const accountSteps: IntegrationStep<IntegrationConfig>[] = [
   {
     id: Steps.ACCOUNT,
-    name: 'Fetch Account Details',
+    name: 'Fetch Account',
     entities: [Entities.ACCOUNT],
     relationships: [],
     dependsOn: [],
-    executionHandler: fetchAccountDetails,
+    executionHandler: fetchAccount,
   },
 ];