+++ src/steps/devices/index.test.ts
@@ -1,0 +1,34 @@
+import { executeStepWithDependencies } from '@jupiterone/integration-sdk-testing';
+import { buildStepTestConfigForStep } from '../../../test/config';
+import { Recording, setupProjectRecording } from '../../../test/recording';
+import { Steps } from '../constants';
+
+// See test/README.md for details
+let recording: Recording;
+afterEach(async () => {
+  await recording.stop();
+});
+
+test('fetch-devices', async () => {
+  recording = setupProjectRecording({
+    directory: __dirname,
+    name: 'fetch-devices',
+  });
+
+  const stepConfig = buildStepTestConfigForStep(Steps.DEVICES);
+  const stepResult = await executeStepWithDependencies(stepConfig);
+  expect(stepResult).toMatchStepMetadata(stepConfig);
+});
+
+test('build-devices-and-users-relationships', async () => {
+  recording = setupProjectRecording({
+    directory: __dirname,
+    name: 'build-devices-and-users-relationships',
+  });
+
+  const stepConfig = buildStepTestConfigForStep(
+    Steps.BUILD_DEVICES_AND_USERS_RELATIONSHIPS,
+  );
+  const stepResult = await executeStepWithDependencies(stepConfig);
+  expect(stepResult).toMatchStepMetadata(stepConfig);
+});