+++ CHANGELOG.md
@@ -7,3 +7,20 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+
+- Ingest new entities
+  - `hexnode_account`
+  - `hexnode_user`
+  - `hexnode_user_group`
+  - `hexnode_device`
+  - `hexnode_device_group`
+- Build new relationships
+  - `hexnode_account_has_device`
+  - `hexnode_account_has_device_group`
+  - `hexnode_account_has_user`
+  - `hexnode_account_has_user_group`
+  - `hexnode_device_group_has_device`
+  - `hexnode_user_group_has_user`
+  - `hexnode_user_has_device`